requirements.txt
discord.py>=2.3.2
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
supabase>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.5.0
python-multipart>=0.0.6
matplotlib>=3.7.0

